refactor: changed grouping of classes into folders
