+++ src/main/java/org/itmo/game/logic/Direction.java
@@ -1,4 +1,4 @@
-package org.itmo.constants;
+package org.itmo.game.logic;
 
 public enum Direction {
     UP,
